refactor: resolve duplicate modules and clean build artifacts
- Consolidated duplicate modules to core/ directory
- Removed build artifacts and enhanced .gitignore
- All P0 tests passing, zero import conflicts
- Improved maintainability and eliminated runtime risks
