+++ .claudedirector/lib/core/hybrid_installation_manager.py
@@ -1,345 +1,445 @@
-#!/usr/bin/env python3
 """
-Hybrid Installation Manager
-
-Provides 58% performance improvement through smart installation strategy:
-- Tier 1: Permanent installation (fastest - 0.46s)
-- Tier 2: Temporary installation (fallback - 1.1s, maintains zero setup)
-- Tier 3: Performance hints after 3+ uses (user choice)
-
-Maintains zero setup principle while enabling performance optimization.
+Hybrid MCP Installation Manager
+Smart detection and fallback system for MCP server installation strategies
 """
 
 import subprocess
 import time
 import json
 from pathlib import Path
-from typing import Dict, Optional, Any
+from typing import Dict, List, Optional, Any
 from dataclasses import dataclass
-import logging
+from datetime import datetime
 
-# Import configuration system
-from .config import get_config
 
-logger = logging.getLogger(__name__)
+@dataclass
+class InstallationCommand:
+    """Represents a single installation command option"""
+
+    type: str  # "permanent" or "temporary"
+    command: str
+    args: List[str]
+    performance_benefit: Optional[str] = None
+    fallback_message: Optional[str] = None
 
 
 @dataclass
 class InstallationResult:
-    """Result of installation attempt"""
+    """Result of attempting to run an MCP server"""
 
     success: bool
-    method: str  # 'permanent', 'temporary', 'cached'
-    duration: float
-    command_available: bool
+    installation_type: str  # "permanent", "temporary", or "failed"
+    command_used: str
+    startup_time: float
     error_message: Optional[str] = None
+    performance_benefit: Optional[str] = None
 
 
 @dataclass
 class UsageStats:
-    """Track usage for performance optimization hints"""
+    """Track usage statistics for performance hints"""
 
-    total_uses: int = 0
-    temporary_install_uses: int = 0
-    last_hint_shown: Optional[float] = None
-    permanent_install_available: bool = False
+    server_name: str
+    temporary_uses: int = 0
+    permanent_uses: int = 0
+    last_hint_shown: Optional[datetime] = None
+    total_startup_time_saved: float = 0.0
 
 
 class HybridInstallationManager:
     """
-    Smart installation manager providing 58% performance improvement
+    Manages hybrid installation strategy for MCP servers
 
     Strategy:
-    1. Check for permanent installation (fastest)
-    2. Fall back to temporary installation (zero setup maintained)
-    3. Track usage and suggest optimizations
+    1. Try permanent installation first (fastest)
+    2. Fallback to temporary installation (maintains zero setup)
+    3. Track usage and suggest optimization when beneficial
     """
 
-    def __init__(self):
-        self.config = get_config()
-        self.usage_file = Path.home() / ".claudedirector" / "usage_stats.json"
+    def __init__(self, config_path: Optional[Path] = None):
+        if config_path is None:
+            config_path = (
+                Path(__file__).parent.parent.parent / "config" / "mcp_servers.yaml"
+            )
+
+        self.config_path = config_path
+        self.config = self._load_config()
         self.usage_stats = self._load_usage_stats()
 
-        # Performance thresholds from configuration
-        self.permanent_install_time = (
-            self.config.get_threshold("performance_degradation_limit") * 10
-        )  # 0.46s
-        self.temporary_install_time = self.permanent_install_time * 2.4  # ~1.1s
-        self.usage_hint_threshold = 3  # Show hint after 3+ uses
-
-    def _load_usage_stats(self) -> UsageStats:
-        """Load usage statistics for optimization hints"""
+    def _load_config(self) -> Dict[str, Any]:
+        """Load MCP server configuration"""
         try:
-            if self.usage_file.exists():
-                with open(self.usage_file, "r") as f:
-                    data = json.load(f)
-                    return UsageStats(**data)
-        except Exception as e:
-            logger.debug(f"Could not load usage stats: {e}")
+            import yaml
 
-        return UsageStats()
+            with open(self.config_path, "r") as f:
+                return yaml.safe_load(f)
+        except ImportError:
+            # Fallback without yaml
+            return self._parse_simple_config()
+        except FileNotFoundError:
+            return self._get_default_config()
 
-    def _save_usage_stats(self):
-        """Save usage statistics"""
-        try:
-            self.usage_file.parent.mkdir(parents=True, exist_ok=True)
-            with open(self.usage_file, "w") as f:
-                json.dump(self.usage_stats.__dict__, f, indent=2)
-        except Exception as e:
-            logger.debug(f"Could not save usage stats: {e}")
+    def _parse_simple_config(self) -> Dict[str, Any]:
+        """Simple config parser for environments without PyYAML"""
+        # Basic fallback - would need to be implemented based on specific needs
+        return self._get_default_config()
 
-    def install_mcp_package(
-        self, package_name: str, timeout: int = 30
-    ) -> InstallationResult:
-        """
-        Install MCP package using hybrid strategy
-
-        Returns installation result with performance metrics
-        """
-        start_time = time.time()
-
-        # Tier 1: Check for permanent installation (fastest)
-        permanent_result = self._check_permanent_installation(package_name)
-        if permanent_result.success:
-            duration = time.time() - start_time
-            logger.info(
-                f"Using permanent installation: {package_name} ({duration:.3f}s)"
-            )
-            return permanent_result
-
-        # Tier 2: Temporary installation (fallback, maintains zero setup)
-        temporary_result = self._temporary_installation(package_name, timeout)
-
-        # Update usage statistics
-        self._update_usage_stats(temporary_result.success)
-
-        # Tier 3: Performance hints (after 3+ uses)
-        if self._should_show_performance_hint():
-            self._show_performance_hint(package_name)
-
-        return temporary_result
-
-    def _check_permanent_installation(self, package_name: str) -> InstallationResult:
-        """Check if package is permanently installed"""
-        start_time = time.time()
-
-        try:
-            # Check if npx can find the package quickly
-            result = subprocess.run(
-                ["npx", "--version"],  # Quick check for npx availability
-                capture_output=True,
-                text=True,
-                timeout=2,
-            )
-
-            if result.returncode != 0:
-                return InstallationResult(
-                    success=False,
-                    method="permanent",
-                    duration=time.time() - start_time,
-                    command_available=False,
-                    error_message="npx not available",
-                )
-
-            # Quick check for installed package
-            check_result = subprocess.run(
-                ["npx", package_name, "--version"],
-                capture_output=True,
-                text=True,
-                timeout=3,
-            )
-
-            duration = time.time() - start_time
-
-            if check_result.returncode == 0:
-                return InstallationResult(
-                    success=True,
-                    method="permanent",
-                    duration=duration,
-                    command_available=True,
-                )
-            else:
-                return InstallationResult(
-                    success=False,
-                    method="permanent",
-                    duration=duration,
-                    command_available=False,
-                    error_message="Package not permanently installed",
-                )
-
-        except subprocess.TimeoutExpired:
-            return InstallationResult(
-                success=False,
-                method="permanent",
-                duration=time.time() - start_time,
-                command_available=False,
-                error_message="Permanent installation check timed out",
-            )
-        except Exception as e:
-            return InstallationResult(
-                success=False,
-                method="permanent",
-                duration=time.time() - start_time,
-                command_available=False,
-                error_message=f"Permanent installation check failed: {e}",
-            )
-
-    def _temporary_installation(
-        self, package_name: str, timeout: int
-    ) -> InstallationResult:
-        """Temporary installation (maintains zero setup principle)"""
-        start_time = time.time()
-
-        try:
-            # Use npx for temporary installation
-            result = subprocess.run(
-                ["npx", "--yes", package_name, "--version"],
-                capture_output=True,
-                text=True,
-                timeout=timeout,
-            )
-
-            duration = time.time() - start_time
-
-            if result.returncode == 0:
-                logger.info(
-                    f"Temporary installation successful: {package_name} ({duration:.3f}s)"
-                )
-                return InstallationResult(
-                    success=True,
-                    method="temporary",
-                    duration=duration,
-                    command_available=True,
-                )
-            else:
-                return InstallationResult(
-                    success=False,
-                    method="temporary",
-                    duration=duration,
-                    command_available=False,
-                    error_message=f"Temporary installation failed: {result.stderr}",
-                )
-
-        except subprocess.TimeoutExpired:
-            return InstallationResult(
-                success=False,
-                method="temporary",
-                duration=time.time() - start_time,
-                command_available=False,
-                error_message=f"Temporary installation timed out after {timeout}s",
-            )
-        except Exception as e:
-            return InstallationResult(
-                success=False,
-                method="temporary",
-                duration=time.time() - start_time,
-                command_available=False,
-                error_message=f"Temporary installation failed: {e}",
-            )
-
-    def _update_usage_stats(self, installation_success: bool):
-        """Update usage statistics for optimization hints"""
-        if installation_success:
-            self.usage_stats.total_uses += 1
-            self.usage_stats.temporary_install_uses += 1
-
-        self._save_usage_stats()
-
-    def _should_show_performance_hint(self) -> bool:
-        """Determine if performance hint should be shown"""
-        # Show hint after 3+ uses, but not more than once per day
-        if self.usage_stats.temporary_install_uses < self.usage_hint_threshold:
-            return False
-
-        # Don't show if already shown recently (24 hours)
-        if self.usage_stats.last_hint_shown:
-            time_since_hint = time.time() - self.usage_stats.last_hint_shown
-            if time_since_hint < 86400:  # 24 hours
-                return False
-
-        return True
-
-    def _show_performance_hint(self, package_name: str):
-        """Show performance optimization hint to user"""
-        improvement_percentage = int(
-            (self.temporary_install_time - self.permanent_install_time)
-            / self.temporary_install_time
-            * 100
-        )
-
-        hint_message = f"""
-💡 Performance Optimization Available
-
-You've used {package_name} {self.usage_stats.temporary_install_uses} times.
-Install permanently for {improvement_percentage}% faster startup:
-
-    npm install -g {package_name}
-
-This reduces startup from ~{self.temporary_install_time:.1f}s to ~{self.permanent_install_time:.1f}s.
-(Optional - zero setup is maintained without this)
-"""
-
-        print(hint_message)
-        logger.info(f"Performance hint shown for {package_name}")
-
-        # Update hint timestamp
-        self.usage_stats.last_hint_shown = time.time()
-        self._save_usage_stats()
-
-    def get_installation_strategy(self, package_name: str) -> str:
-        """Get recommended installation strategy message"""
-        permanent_result = self._check_permanent_installation(package_name)
-
-        if permanent_result.success:
-            improvement = int(
-                (self.temporary_install_time - permanent_result.duration)
-                / self.temporary_install_time
-                * 100
-            )
-            return f"⚡ Using optimized MCP server: {package_name} ({improvement}% faster startup)"
-        else:
-            return f"🔧 Installing MCP enhancement: {package_name} (automatic setup)"
-
-    def get_performance_stats(self) -> Dict[str, Any]:
-        """Get performance statistics for reporting"""
+    def _get_default_config(self) -> Dict[str, Any]:
+        """Default configuration when config file unavailable"""
         return {
-            "total_uses": self.usage_stats.total_uses,
-            "temporary_uses": self.usage_stats.temporary_install_uses,
-            "permanent_install_time": self.permanent_install_time,
-            "temporary_install_time": self.temporary_install_time,
-            "performance_improvement": f"{int((self.temporary_install_time - self.permanent_install_time) / self.temporary_install_time * 100)}%",
+            "servers": {
+                "sequential": {
+                    "installation_strategy": "hybrid",
+                    "commands": [
+                        {
+                            "type": "permanent",
+                            "command": "sequential-mcp-server",
+                            "args": [],
+                            "performance_benefit": "58% faster startup",
+                        },
+                        {
+                            "type": "temporary",
+                            "command": "npx",
+                            "args": ["-y", "@sequential/mcp-server"],
+                            "fallback_message": "Installing MCP enhancement: sequential (systematic_analysis)",
+                        },
+                    ],
+                }
+            },
+            "hybrid_installation": {
+                "track_installation_performance": True,
+                "performance_hint_threshold": 3,
+                "messages": {
+                    "permanent_found": "⚡ Using optimized MCP server: {server} ({benefit})",
+                    "temporary_installing": "🔧 Installing MCP enhancement: {server} ({capability})",
+                    "performance_hint": "💡 Install {server} permanently for {benefit}? Use: npm install -g {package}",
+                },
+            },
         }
 
+    def _load_usage_stats(self) -> Dict[str, UsageStats]:
+        """Load usage statistics for performance tracking"""
+        stats_file = (
+            Path(__file__).parent.parent.parent / "data" / "mcp_usage_stats.json"
+        )
+        stats = {}
 
-# Singleton instance for global use
-_hybrid_manager_instance: Optional[HybridInstallationManager] = None
+        try:
+            if stats_file.exists():
+                with open(stats_file, "r") as f:
+                    data = json.load(f)
+                    for server_name, stats_data in data.items():
+                        stats[server_name] = UsageStats(
+                            server_name=server_name,
+                            temporary_uses=stats_data.get("temporary_uses", 0),
+                            permanent_uses=stats_data.get("permanent_uses", 0),
+                            last_hint_shown=(
+                                datetime.fromisoformat(stats_data["last_hint_shown"])
+                                if stats_data.get("last_hint_shown")
+                                else None
+                            ),
+                            total_startup_time_saved=stats_data.get(
+                                "total_startup_time_saved", 0.0
+                            ),
+                        )
+        except (FileNotFoundError, json.JSONDecodeError, KeyError):
+            pass  # Start with empty stats
 
+        return stats
 
-def get_hybrid_manager() -> HybridInstallationManager:
-    """Get global hybrid installation manager instance"""
-    global _hybrid_manager_instance
-    if _hybrid_manager_instance is None:
-        _hybrid_manager_instance = HybridInstallationManager()
-    return _hybrid_manager_instance
+    def _save_usage_stats(self):
+        """Save usage statistics to disk"""
+        stats_file = (
+            Path(__file__).parent.parent.parent / "data" / "mcp_usage_stats.json"
+        )
+        stats_file.parent.mkdir(exist_ok=True)
 
+        data = {}
+        for server_name, stats in self.usage_stats.items():
+            data[server_name] = {
+                "temporary_uses": stats.temporary_uses,
+                "permanent_uses": stats.permanent_uses,
+                "last_hint_shown": (
+                    stats.last_hint_shown.isoformat() if stats.last_hint_shown else None
+                ),
+                "total_startup_time_saved": stats.total_startup_time_saved,
+            }
 
-# Convenience functions for backwards compatibility
-def install_mcp_package(package_name: str, timeout: int = 30) -> InstallationResult:
-    """Install MCP package using hybrid strategy"""
-    return get_hybrid_manager().install_mcp_package(package_name, timeout)
+        try:
+            with open(stats_file, "w") as f:
+                json.dump(data, f, indent=2)
+        except Exception:
+            pass  # Graceful failure - don't break functionality for stats
 
+    def check_command_availability(self, command: str) -> bool:
+        """Check if a command is available in the system PATH"""
+        try:
+            # For python commands, use a more reliable check
+            if command in ["python", "python3"]:
+                result = subprocess.run(
+                    [command, "--version"], capture_output=True, text=True, timeout=3
+                )
+                return result.returncode == 0
 
-def get_installation_strategy_message(package_name: str) -> str:
-    """Get installation strategy message for transparency"""
-    return get_hybrid_manager().get_installation_strategy(package_name)
+            # For other commands, try which first
+            result = subprocess.run(
+                ["which", command], capture_output=True, text=True, timeout=3
+            )
+            if result.returncode == 0:
+                return True
 
+            # Fallback to version check
+            result = subprocess.run(
+                [command, "--version"], capture_output=True, text=True, timeout=3
+            )
+            return result.returncode == 0
+        except (
+            subprocess.TimeoutExpired,
+            FileNotFoundError,
+            subprocess.SubprocessError,
+        ):
+            return False
 
-# Export main classes and functions
-__all__ = [
-    "HybridInstallationManager",
-    "InstallationResult",
-    "UsageStats",
-    "get_hybrid_manager",
-    "install_mcp_package",
-    "get_installation_strategy_message",
-]
+    def try_installation_command(
+        self, cmd: InstallationCommand, timeout: int = 10
+    ) -> InstallationResult:
+        """Try to execute a single installation command"""
+        start_time = time.time()
+
+        try:
+            # For permanent commands, first check if they're available
+            if cmd.type == "permanent" and not self.check_command_availability(
+                cmd.command
+            ):
+                return InstallationResult(
+                    success=False,
+                    installation_type="permanent",
+                    command_used=f"{cmd.command} {' '.join(cmd.args)}",
+                    startup_time=time.time() - start_time,
+                    error_message=f"Command '{cmd.command}' not found in PATH",
+                )
+
+            # Try to execute the command
+            full_command = (
+                [cmd.command] + cmd.args + ["--version"]
+            )  # Just test availability
+            result = subprocess.run(
+                full_command, capture_output=True, text=True, timeout=timeout
+            )
+
+            startup_time = time.time() - start_time
+
+            if result.returncode == 0:
+                return InstallationResult(
+                    success=True,
+                    installation_type=cmd.type,
+                    command_used=f"{cmd.command} {' '.join(cmd.args)}",
+                    startup_time=startup_time,
+                    performance_benefit=cmd.performance_benefit,
+                )
+            else:
+                return InstallationResult(
+                    success=False,
+                    installation_type=cmd.type,
+                    command_used=f"{cmd.command} {' '.join(cmd.args)}",
+                    startup_time=startup_time,
+                    error_message=result.stderr or "Command failed",
+                )
+
+        except subprocess.TimeoutExpired:
+            return InstallationResult(
+                success=False,
+                installation_type=cmd.type,
+                command_used=f"{cmd.command} {' '.join(cmd.args)}",
+                startup_time=time.time() - start_time,
+                error_message="Command timed out",
+            )
+        except Exception as e:
+            return InstallationResult(
+                success=False,
+                installation_type=cmd.type,
+                command_used=f"{cmd.command} {' '.join(cmd.args)}",
+                startup_time=time.time() - start_time,
+                error_message=str(e),
+            )
+
+    def get_installation_commands(self, server_name: str) -> List[InstallationCommand]:
+        """Get ordered list of installation commands for a server"""
+        server_config = self.config.get("servers", {}).get(server_name, {})
+
+        if server_config.get("installation_strategy") == "hybrid":
+            commands = []
+            for cmd_config in server_config.get("commands", []):
+                commands.append(
+                    InstallationCommand(
+                        type=cmd_config.get("type", "temporary"),
+                        command=cmd_config.get("command", ""),
+                        args=cmd_config.get("args", []),
+                        performance_benefit=cmd_config.get("performance_benefit"),
+                        fallback_message=cmd_config.get("fallback_message"),
+                    )
+                )
+            return commands
+        else:
+            # Legacy configuration - convert to hybrid
+            return [
+                InstallationCommand(
+                    type="temporary",
+                    command=server_config.get("command", "npx"),
+                    args=server_config.get(
+                        "args", ["-y", f"@{server_name}/mcp-server"]
+                    ),
+                    fallback_message=f"Installing MCP enhancement: {server_name}",
+                )
+            ]
+
+    def attempt_server_startup(self, server_name: str) -> InstallationResult:
+        """
+        Attempt to start an MCP server using hybrid installation strategy
+
+        Returns the result of the first successful installation method
+        """
+        commands = self.get_installation_commands(server_name)
+
+        for cmd in commands:
+            result = self.try_installation_command(cmd)
+
+            if result.success:
+                # Track usage statistics
+                self._track_usage(
+                    server_name, result.installation_type, result.startup_time
+                )
+
+                # Show appropriate message
+                self._show_installation_message(server_name, result)
+
+                # Check if we should show performance hint
+                self._check_performance_hint(server_name, result.installation_type)
+
+                return result
+
+        # All commands failed
+        return InstallationResult(
+            success=False,
+            installation_type="failed",
+            command_used="all commands failed",
+            startup_time=0.0,
+            error_message="No installation method succeeded",
+        )
+
+    def _track_usage(
+        self, server_name: str, installation_type: str, startup_time: float
+    ):
+        """Track usage statistics for performance optimization hints"""
+        if not self.config.get("hybrid_installation", {}).get(
+            "track_installation_performance", True
+        ):
+            return
+
+        if server_name not in self.usage_stats:
+            self.usage_stats[server_name] = UsageStats(server_name=server_name)
+
+        stats = self.usage_stats[server_name]
+
+        if installation_type == "permanent":
+            stats.permanent_uses += 1
+        elif installation_type == "temporary":
+            stats.temporary_uses += 1
+            # Track potential time savings
+            stats.total_startup_time_saved += max(
+                0, startup_time - 0.46
+            )  # Assume permanent would be ~0.46s
+
+        self._save_usage_stats()
+
+    def _show_installation_message(self, server_name: str, result: InstallationResult):
+        """Show appropriate message based on installation type"""
+        messages = self.config.get("hybrid_installation", {}).get("messages", {})
+
+        if result.installation_type == "permanent" and result.performance_benefit:
+            msg = messages.get(
+                "permanent_found", "⚡ Using optimized MCP server: {server} ({benefit})"
+            )
+            print(msg.format(server=server_name, benefit=result.performance_benefit))
+        elif result.installation_type == "temporary":
+            # This message is handled by the transparency bridge for consistency
+            pass
+
+    def _check_performance_hint(self, server_name: str, installation_type: str):
+        """Check if we should show a performance optimization hint"""
+        if installation_type != "temporary":
+            return
+
+        stats = self.usage_stats.get(server_name)
+        if not stats:
+            return
+
+        threshold = self.config.get("hybrid_installation", {}).get(
+            "performance_hint_threshold", 3
+        )
+
+        # Show hint after threshold temporary uses
+        if (
+            stats.temporary_uses >= threshold
+            and stats.permanent_uses == 0
+            and (
+                not stats.last_hint_shown
+                or (datetime.now() - stats.last_hint_shown).days >= 7
+            )
+        ):  # Don't spam hints
+
+            self._show_performance_hint(server_name, stats)
+            stats.last_hint_shown = datetime.now()
+            self._save_usage_stats()
+
+    def _show_performance_hint(self, server_name: str, stats: UsageStats):
+        """Show performance optimization hint to user"""
+        messages = self.config.get("hybrid_installation", {}).get("messages", {})
+        commands = self.get_installation_commands(server_name)
+
+        permanent_cmd = next((cmd for cmd in commands if cmd.type == "permanent"), None)
+        if not permanent_cmd:
+            return
+
+        # Calculate package name from command
+        package_name = f"@{server_name}/mcp-server"  # Default guess
+
+        hint_msg = messages.get(
+            "performance_hint",
+            "💡 Install {server} permanently for {benefit}? Use: npm install -g {package}",
+        )
+
+        print(
+            f"\n{hint_msg.format(server=server_name, benefit=permanent_cmd.performance_benefit or '58% faster startup', package=package_name)}"
+        )
+        print(
+            f"   You've used {server_name} {stats.temporary_uses} times - permanent installation saves ~{stats.total_startup_time_saved:.1f}s total"
+        )
+        print(
+            "   Zero setup is maintained - this optimization is completely optional\n"
+        )
+
+    def get_server_status(self, server_name: str) -> Dict[str, Any]:
+        """Get comprehensive status for a server"""
+        commands = self.get_installation_commands(server_name)
+        stats = self.usage_stats.get(server_name, UsageStats(server_name=server_name))
+
+        permanent_available = False
+        if commands:
+            permanent_cmd = next(
+                (cmd for cmd in commands if cmd.type == "permanent"), None
+            )
+            if permanent_cmd:
+                permanent_available = self.check_command_availability(
+                    permanent_cmd.command
+                )
+
+        return {
+            "server_name": server_name,
+            "permanent_available": permanent_available,
+            "temporary_uses": stats.temporary_uses,
+            "permanent_uses": stats.permanent_uses,
+            "total_time_saved": stats.total_startup_time_saved,
+            "installation_strategy": "hybrid" if len(commands) > 1 else "legacy",
+            "commands_available": len(commands),
+        }
