--- .claudedirector/test_results/test_results_20250823_232220.json
@@ -1,138 +1,0 @@
-{
-  "timestamp": "20250823_232220",
-  "environment": "pre_push",
-  "duration": 25.013061046600342,
-  "overall_success": true,
-  "suites": [
-    {
-      "name": "quality_gates",
-      "description": "Code quality and security validation",
-      "status": "passed",
-      "duration": 1.4015097618103027,
-      "blocking": false,
-      "tests": [
-        {
-          "name": "Black Formatting",
-          "status": "passed",
-          "duration": 0.32710719108581543,
-          "critical": false,
-          "error": null
-        },
-        {
-          "name": "Security Scan",
-          "status": "passed",
-          "duration": 0.05262899398803711,
-          "critical": false,
-          "error": null
-        },
-        {
-          "name": "Flake8 Linting",
-          "status": "failed",
-          "duration": 1.0215718746185303,
-          "critical": false,
-          "error": ""
-        }
-      ]
-    },
-    {
-      "name": "p0_blocking",
-      "description": "Critical P0 tests that must pass for CI success",
-      "status": "passed",
-      "duration": 23.611209869384766,
-      "blocking": true,
-      "tests": [
-        {
-          "name": "MCP Transparency P0",
-          "status": "passed",
-          "duration": 0.04545474052429199,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Conversation Tracking P0",
-          "status": "passed",
-          "duration": 0.15392780303955078,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Conversation Quality P0",
-          "status": "passed",
-          "duration": 0.12224531173706055,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Configuration Persistence P0",
-          "status": "passed",
-          "duration": 0.043992042541503906,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "ROI Tracking P0",
-          "status": "passed",
-          "duration": 0.046714067459106445,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Security P0",
-          "status": "passed",
-          "duration": 0.055066823959350586,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Error Recovery P0",
-          "status": "passed",
-          "duration": 0.041532039642333984,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Performance P0",
-          "status": "passed",
-          "duration": 22.835235834121704,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Persona Consistency P0",
-          "status": "passed",
-          "duration": 0.05814719200134277,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Framework Attribution System P0",
-          "status": "passed",
-          "duration": 0.07912611961364746,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Context Switching P0",
-          "status": "passed",
-          "duration": 0.048177242279052734,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "MCP Integration P0",
-          "status": "passed",
-          "duration": 0.04037594795227051,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Persona Strategic Thinking P0",
-          "status": "passed",
-          "duration": 0.04012584686279297,
-          "critical": true,
-          "error": null
-        }
-      ]
-    }
-  ]
-}
