--- .claudedirector/test_results/test_results_20250823_234417.json
@@ -1,138 +1,0 @@
-{
-  "timestamp": "20250823_234417",
-  "environment": "pre_push",
-  "duration": 24.477270126342773,
-  "overall_success": true,
-  "suites": [
-    {
-      "name": "quality_gates",
-      "description": "Code quality and security validation",
-      "status": "passed",
-      "duration": 1.3781111240386963,
-      "blocking": false,
-      "tests": [
-        {
-          "name": "Black Formatting",
-          "status": "passed",
-          "duration": 0.3125631809234619,
-          "critical": false,
-          "error": null
-        },
-        {
-          "name": "Security Scan",
-          "status": "passed",
-          "duration": 0.051084041595458984,
-          "critical": false,
-          "error": null
-        },
-        {
-          "name": "Flake8 Linting",
-          "status": "passed",
-          "duration": 1.014232873916626,
-          "critical": false,
-          "error": null
-        }
-      ]
-    },
-    {
-      "name": "p0_blocking",
-      "description": "Critical P0 tests that must pass for CI success",
-      "status": "passed",
-      "duration": 23.098811149597168,
-      "blocking": true,
-      "tests": [
-        {
-          "name": "MCP Transparency P0",
-          "status": "passed",
-          "duration": 0.04293513298034668,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Conversation Tracking P0",
-          "status": "passed",
-          "duration": 0.15328693389892578,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Conversation Quality P0",
-          "status": "passed",
-          "duration": 0.1231689453125,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Configuration Persistence P0",
-          "status": "passed",
-          "duration": 0.048152923583984375,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "ROI Tracking P0",
-          "status": "passed",
-          "duration": 0.04371190071105957,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Security P0",
-          "status": "passed",
-          "duration": 0.052555084228515625,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Error Recovery P0",
-          "status": "passed",
-          "duration": 0.04225897789001465,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Performance P0",
-          "status": "passed",
-          "duration": 22.362540245056152,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Persona Consistency P0",
-          "status": "passed",
-          "duration": 0.048188209533691406,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Framework Attribution System P0",
-          "status": "passed",
-          "duration": 0.05378913879394531,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Context Switching P0",
-          "status": "passed",
-          "duration": 0.04636883735656738,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "MCP Integration P0",
-          "status": "passed",
-          "duration": 0.039212703704833984,
-          "critical": true,
-          "error": null
-        },
-        {
-          "name": "Persona Strategic Thinking P0",
-          "status": "passed",
-          "duration": 0.04160284996032715,
-          "critical": true,
-          "error": null
-        }
-      ]
-    }
-  ]
-}
